requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.17.1
gunicorn==25.3.0
mysqlclient==2.2.8
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
